Moved endpoints to zooRoutes.js
